requirements.txt
annotated-types==0.6.0
anyio==4.3.0
asgiref==3.8.1
attrs==23.2.0
click==8.1.7
Django==4.1.13
django-cors-headers==4.3.1
django-restframework==0.0.1
djangorestframework==3.15.1
djongo==1.3.6
dnspython==2.6.1
drf-spectacular==0.27.2
drf-yasg==1.21.7
exceptiongroup==1.2.0
fastapi==0.110.1
generics==6.0.0
h11==0.14.0
idna==3.6
inflection==0.5.1
jsonschema==4.22.0
jsonschema-specifications==2023.12.1
line-profiler==4.1.2
mongoengine==0.28.2
npx==0.1.2
numpy==1.26.4
packaging==24.0
pydantic==2.6.4
pydantic_core==2.16.3
pymongo==4.6.3
pytz==2024.1
PyYAML==6.0.1
referencing==0.35.1
rpds-py==0.18.1
sniffio==1.3.1
sqlparse==0.2.4
starlette==0.37.2
typing_extensions==4.10.0
uritemplate==4.1.1
uvicorn==0.29.0
